requirements.txt
# Core
pandas>=2.0.0
pyarrow>=14.0.0
numpy>=1.26.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# API server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# HTTP client (extraction script)
requests>=2.31.0
urllib3>=2.0.0

# Config
python-dotenv>=1.0.0

# Progress bars
tqdm>=4.66.0

# Experiment tracking (optional but recommended)
mlflow>=2.10.0


dvc>=3.0.0